Revert "wrong commit which should be for DAOS-6135"
This reverts commit db70f1e146bd14c29cc7450eead6c839df218d0c.
