+++ src/client/java/daos-java/pom.xml
@@ -43,11 +43,6 @@
       <version>3.11.4</version>
     </dependency>
     <dependency>
-      <groupId>io.netty</groupId>
-      <artifactId>netty-buffer</artifactId>
-      <version>4.1.45.Final</version>
-    </dependency>
-    <dependency>
       <groupId>com.google.code.findbugs</groupId>
       <artifactId>jsr305</artifactId>
     </dependency>
@@ -68,7 +63,6 @@
         <configuration>
           <javahClassNames>
             <javahClassName>io.daos.dfs.DaosFsClient</javahClassName>
-            <javahClassName>io.daos.DaosClient</javahClassName>
           </javahClassNames>
           <javahOutputDirectory>${project.basedir}/src/main/native/include</javahOutputDirectory>
         </configuration>
@@ -122,37 +116,6 @@
                 </exec>
                 <echo message="Java home is ${java-home}" />
                 <delete file="${project.basedir}/src/main/native/include/io_daos_dfs_DaosFsClient_DaosFsClientBuilder.h"/>
-                <delete file="${project.basedir}/src/main/native/include/io_daos_DaosClient_DaosClientBuilder.h"/>
-                <exec dir="${native.build.path}"
-                      failonerror="true"
-                      executable="gcc">
-                  <arg value="-v"/>
-                  <arg value="-I/usr/local/"/>
-                  <arg value="-I${java-home}/include/"/>
-                  <arg value="-I${java-home}/include/linux/"/>
-                  <arg value="-I${daos.install.path}/include/"/>
-                  <arg value="-I${project.basedir}/src/main/native/include/"/>
-                  <arg value="${project.basedir}/src/main/native/daos_jni_common.c"/>
-                  <arg value="-o${native.build.path}/daos_jni_common.o"/>
-                  <arg value="-c"/>
-                  <arg value="-fPIC"/>
-                  <arg value="--std=gnu99"/>
-                </exec>
-                <exec dir="${native.build.path}"
-                      failonerror="true"
-                      executable="gcc">
-                  <arg value="-v"/>
-                  <arg value="-I/usr/local/"/>
-                  <arg value="-I${java-home}/include/"/>
-                  <arg value="-I${java-home}/include/linux/"/>
-                  <arg value="-I${daos.install.path}/include/"/>
-                  <arg value="-I${project.basedir}/src/main/native/include/"/>
-                  <arg value="${project.basedir}/src/main/native/io_daos_DaosClient.c"/>
-                  <arg value="-o${native.build.path}/io_daos_DaosClient.o"/>
-                  <arg value="-c"/>
-                  <arg value="-fPIC"/>
-                  <arg value="--std=gnu99"/>
-                </exec>
                 <exec dir="${native.build.path}"
                       failonerror="true"
                       executable="gcc">
@@ -190,10 +153,8 @@
                       executable="gcc">
                   <arg value="-shared"/>
                   <arg value="-o${native.build.path}/libdaos-jni.so"/>
-                  <arg value="${native.build.path}/daos_jni_common.o"/>
                   <arg value="${native.build.path}/io_daos_dfs_DaosFsClient.o"/>
                   <arg value="${native.build.path}/DunsAttribute.pb-c.o"/>
-                  <arg value="${native.build.path}/io_daos_DaosClient.o"/>
                   <arg value="-L${daos.install.path}/lib"/>
                   <arg value="-L${daos.install.path}/lib64"/>
                   <arg value="-L${daos.install.path}/prereq/dev/protobufc/lib"/>
