+++ src/client/java/daos-java/src/main/java/io/daos/dfs/DaosFile.java
@@ -28,7 +28,6 @@ import java.nio.ByteBuffer;
 
 import javax.annotation.concurrent.NotThreadSafe;
 
-import io.daos.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
