+++ src/client/java/daos-java/src/main/java/io/daos/dfs/ErrorCode.java
@@ -21,22 +21,25 @@
  * portions thereof marked with this legend must also reproduce the markings.
  */
 
-package io.daos;
-
-import java.io.Closeable;
-import java.io.IOException;
+package io.daos.dfs;
 
 /**
- * A interface extended from {@link Closeable} to add one more method {@link #forceClose()}.
- * When it's called, resource cleanup or disconnection should be done immediately no matter
- * what circumstance it is. It's usually called at JVM shutdown.
+ * error code to message.
  */
-public interface ForceCloseable extends Closeable {
+public class ErrorCode {
+  private final int code;
+  private final String msg;
 
-  /**
-   * close at any circumstance.
-   *
-   * @throws IOException
-   */
-  void forceClose() throws IOException;
+  public ErrorCode(int code, String msg) {
+    this.code = code;
+    this.msg = msg;
+  }
+
+  public int getCode() {
+    return code;
+  }
+
+  public String getMsg() {
+    return msg;
+  }
 }
