+++ src/client/java/daos-java/src/main/native/include/daos_jni_common.h
@@ -56,89 +56,5 @@ static const int CUSTOM_ERR3 = -1000003;
 static const int CUSTOM_ERR4 = -1000004;
 /* invalid argument in UNS */
 static const int CUSTOM_ERR5 = -1000005;
-/* invalid argument in object */
-static const int CUSTOM_ERR6 = -1000006;
-
-static jclass daos_io_exception_class;
-
-static jmethodID new_exception_msg;
-static jmethodID new_exception_cause;
-static jmethodID new_exception_msg_code_msg;
-static jmethodID new_exception_msg_code_cause;
-
-static int ERROR_PATH_LEN = 256;
-static int ERROR_NOT_EXIST = 2;
-static int ERROR_LOOKUP_MAX_RETRIES = 100;
-
-static uint8_t KEY_LIST_CODE_EMPTY = (uint8_t)0;
-static uint8_t KEY_LIST_CODE_IN_USE = (uint8_t)1;
-static uint8_t KEY_LIST_CODE_ANCHOR_END = (uint8_t)2;
-static uint8_t KEY_LIST_CODE_KEY2BIG = (uint8_t)3;
-static uint8_t KEY_LIST_CODE_REACH_LIMIT = (uint8_t)4;
-
-/**
- * utility function to throw Java exception.
- *
- * \param[in]	env		JNI environment
- * \param[in]	msg		error message provided by caller
- * \param[in]	error_code	non-zero return code of DFS function or
- * 				customized error code
- * \param[in]	release_msg	is \a msg needed to be released, (0 or 1)
- * \param[in]	posix_error	is \a error_code posix error,
- * 				1 for true, 0 for false
- *
- * \return	return code of Throw function of \a env
- */
-int
-throw_exception_base(JNIEnv *env, char *msg, int error_code,
-			int release_msg, int posix_error);
-
-/**
- * throw Java exception with dynamically constructed message for posix error.
- *
- * \param[in]	env		JNI environment
- * \param[in]	msg		error message provided by caller
- * \param[in]	error_code	non-zero return code of DFS function
- *
- * \return	return code of throw_exception_base
- */
-int
-throw_exception(JNIEnv *env, char *msg, int error_code);
-
-/**
- * throw Java exception with dynamically constructed message for object error.
- *
- * \param[in]	env		JNI environment
- * \param[in]	msg		error message provided by caller
- * \param[in]	error_code	non-zero return code of DFS function
- *
- * \return	return code of throw_exception_base
- */
-int
-throw_exception_object(JNIEnv *env, char *msg, int error_code);
-
-/**
- * throw Java exception with constant message for posix error.
- *
- * \param[in]	env		JNI environment
- * \param[in]	msg		error message provided by caller
- * \param[in]	error_code	non-zero return code of DFS function
- *
- * \return	return code of throw_exception_base
- */
-int
-throw_exception_const_msg(JNIEnv *env, char *msg, int error_code);
-
-/**
- * throw Java exception with constant message for object error.
- *
- * \param[in]	env		JNI environment
- * \param[in]	msg		error message provided by caller
- * \param[in]	error_code	non-zero return code of DFS function
- *
- * \return	return code of throw_exception_base
- */
-int
-throw_exception_const_msg_object(JNIEnv *env, char *msg, int error_code);
 
 #endif
