+++ src/client/java/daos-java/src/test/java/io/daos/dfs/DaosFileIT.java
@@ -1,7 +1,6 @@
 package io.daos.dfs;
 
 import com.sun.security.auth.module.UnixSystem;
-import io.daos.*;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -22,8 +21,8 @@ public class DaosFileIT {
 
   @BeforeClass
   public static void setup() throws Exception {
-    poolId = DaosTestBase.getPoolId();
-    contId = DaosTestBase.getContId();
+    poolId = System.getProperty("pool_id", DaosFsClientTestBase.DEFAULT_POOL_ID);
+    contId = System.getProperty("cont_id", DaosFsClientTestBase.DEFAULT_CONT_ID);
 
     client = DaosFsClientTestBase.prepareFs(poolId, contId);
   }
@@ -335,7 +334,7 @@ public class DaosFileIT {
   @AfterClass
   public static void teardown() throws Exception {
     if (client != null) {
-      client.close();
+      client.disconnect();
     }
   }
 }
