+++ src/client/java/daos-java/src/test/java/io/daos/dfs/DaosFileMultiThreadsIT.java
@@ -1,9 +1,5 @@
 package io.daos.dfs;
 
-import io.daos.Constants;
-import io.daos.DaosIOException;
-import io.daos.DaosObjectType;
-import io.daos.DaosTestBase;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -22,8 +18,8 @@ public class DaosFileMultiThreadsIT {
 
   @BeforeClass
   public static void setup() throws Exception {
-    poolId = DaosTestBase.getPoolId();
-    contId = DaosTestBase.getContId();
+    poolId = System.getProperty("pool_id", DaosFsClientTestBase.DEFAULT_POOL_ID);
+    contId = System.getProperty("cont_id", DaosFsClientTestBase.DEFAULT_CONT_ID);
 
     client = DaosFsClientTestBase.prepareFs(poolId, contId);
   }
@@ -100,7 +96,7 @@ public class DaosFileMultiThreadsIT {
   @AfterClass
   public static void teardown() throws Exception {
     if (client != null) {
-      client.close();
+      client.disconnect();
     }
   }
