+++ src/client/java/daos-java/src/test/java/io/daos/dfs/DaosFsClientIT.java
@@ -1,6 +1,5 @@
 package io.daos.dfs;
 
-import io.daos.DaosTestBase;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -14,8 +13,8 @@ public class DaosFsClientIT {
 
   @BeforeClass
   public static void setup() throws Exception {
-    poolId = DaosTestBase.getPoolId();
-    contId = DaosTestBase.getContId();
+    poolId = System.getProperty("pool_id", DaosFsClientTestBase.DEFAULT_POOL_ID);
+    contId = System.getProperty("cont_id", DaosFsClientTestBase.DEFAULT_CONT_ID);
   }
 
   @Test
@@ -28,7 +27,7 @@ public class DaosFsClientIT {
       Assert.assertTrue(client != null);
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -43,7 +42,7 @@ public class DaosFsClientIT {
       Assert.assertTrue(client != null);
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -72,7 +71,7 @@ public class DaosFsClientIT {
       thread.join();
       Assert.assertEquals(client, client2[0]);
     } finally {
-      client.close();
+      client.disconnect();
     }
   }
 
@@ -87,7 +86,7 @@ public class DaosFsClientIT {
       client.delete("/ddddddd/zyx", true);
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -104,7 +103,7 @@ public class DaosFsClientIT {
       client.mkdir("/mkdirs/1", true);
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -121,7 +120,7 @@ public class DaosFsClientIT {
       client.mkdir("/mkdir/1", false);
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -137,7 +136,7 @@ public class DaosFsClientIT {
       client.move(0, fileName, 0, "destFile");
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -153,7 +152,7 @@ public class DaosFsClientIT {
       client.move(0, fileName, 0, "/destFile");
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
       }
     }
   }
@@ -179,7 +178,55 @@ public class DaosFsClientIT {
       Assert.assertTrue(client.getFile(destDir, destFileName).exists());
     } finally {
       if (client != null) {
-        client.close();
+        client.disconnect();
+      }
+    }
+  }
+
+  @Test
+  public void testFsClientReferenceOne() throws Exception {
+    DaosFsClient.DaosFsClientBuilder builder = new DaosFsClient.DaosFsClientBuilder();
+    builder.poolId(poolId).containerId(contId);
+    DaosFsClient client = null;
+    try {
+      client = builder.build();
+      Assert.assertEquals(1, client.getRefCnt());
+    } finally {
+      if (client != null) {
+        client.disconnect();
+        Assert.assertEquals(0, client.getRefCnt());
+      }
+    }
+    Exception ee = null;
+    try {
+      client.incrementRef();
+    } catch (Exception e) {
+      ee = e;
+    }
+    Assert.assertTrue(ee instanceof IllegalStateException);
+  }
+
+  @Test
+  public void testFsClientReferenceMore() throws Exception {
+    DaosFsClient.DaosFsClientBuilder builder = new DaosFsClient.DaosFsClientBuilder();
+    builder.poolId(poolId).containerId(contId);
+    DaosFsClient client = null;
+    int cnt = 0;
+    try {
+      client = builder.build();
+      cnt = client.getRefCnt();
+      builder.build();
+      Assert.assertEquals(cnt + 1, client.getRefCnt());
+      client.disconnect();
+      client.incrementRef();
+      Assert.assertEquals(cnt + 1, client.getRefCnt());
+      client.decrementRef();
+    } catch (Exception e) {
+      e.printStackTrace();
+    } finally {
+      if (client != null) {
+        client.disconnect();
+        Assert.assertEquals(cnt - 1, client.getRefCnt());
       }
     }
   }
