+++ src/client/java/hadoop-daos/src/main/java/io/daos/fs/hadoop/DaosFileSystem.java
@@ -33,7 +33,6 @@ import java.util.Set;
 
 import com.google.common.collect.Lists;
 
-import io.daos.*;
 import io.daos.dfs.*;
 
 import org.apache.commons.lang.StringEscapeUtils;
@@ -68,7 +67,7 @@ import org.slf4j.LoggerFactory;
  * <tbody>
  * <tr>
  *   <td>{@value io.daos.fs.hadoop.Constants#DAOS_SERVER_GROUP}</td>
- *   <td>{@value io.daos.Constants#POOL_DEFAULT_SERVER_GROUP}</td>
+ *   <td>{@value io.daos.dfs.Constants#POOL_DEFAULT_SERVER_GROUP}</td>
  *   <td></td>
  *   <td>false</td>
  *   <td>daos server group name</td>
@@ -82,11 +81,11 @@ import org.slf4j.LoggerFactory;
  * </tr>
  * <tr>
  *   <td>{@value io.daos.fs.hadoop.Constants#DAOS_POOL_FLAGS}</td>
- *   <td>{@value io.daos.Constants#ACCESS_FLAG_POOL_READWRITE}</td>
+ *   <td>{@value io.daos.dfs.Constants#ACCESS_FLAG_POOL_READWRITE}</td>
  *   <td>
- *       {@value io.daos.Constants#ACCESS_FLAG_POOL_READONLY},
- *       {@value io.daos.Constants#ACCESS_FLAG_POOL_READWRITE},
- *       {@value io.daos.Constants#ACCESS_FLAG_POOL_EXECUTE}
+ *       {@value io.daos.dfs.Constants#ACCESS_FLAG_POOL_READONLY},
+ *       {@value io.daos.dfs.Constants#ACCESS_FLAG_POOL_READWRITE},
+ *       {@value io.daos.dfs.Constants#ACCESS_FLAG_POOL_EXECUTE}
  *   </td>
  *   <td>false</td>
  *   <td>pool access flags</td>
@@ -170,8 +169,8 @@ public class DaosFileSystem extends FileSystem {
   private String workPath;
 
   static {
-    if (ShutdownHookManager.removeHook(DaosClient.FINALIZER)) {
-      org.apache.hadoop.util.ShutdownHookManager.get().addShutdownHook(DaosClient.FINALIZER, 0);
+    if (ShutdownHookManager.removeHook(DaosFsClient.FINALIZER)) {
+      org.apache.hadoop.util.ShutdownHookManager.get().addShutdownHook(DaosFsClient.FINALIZER, 0);
       if (LOG.isDebugEnabled()) {
         LOG.debug("daos finalizer relocated to hadoop ShutdownHookManager");
       }
@@ -248,7 +247,7 @@ public class DaosFileSystem extends FileSystem {
       if (file.exists()) {
         try {
           info = DaosUns.getAccessInfo(file.getAbsolutePath(), Constants.UNS_ATTR_NAME_HADOOP,
-            io.daos.Constants.UNS_ATTR_VALUE_MAX_LEN_DEFAULT, false);
+            io.daos.dfs.Constants.UNS_ATTR_VALUE_MAX_LEN_DEFAULT, false);
           if (info != null) {
             break;
           }
@@ -928,7 +927,7 @@ public class DaosFileSystem extends FileSystem {
     }
     super.close();
     if (daos != null) {
-      daos.close();
+      daos.disconnect();
     }
   }
