+++ src/client/java/hadoop-daos/src/main/java/io/daos/fs/hadoop/HadoopDaosUtils.java
@@ -29,7 +29,7 @@ import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.NotDirectoryException;
 
-import io.daos.DaosIOException;
+import io.daos.dfs.DaosIOException;
 
 import org.apache.hadoop.fs.ParentNotDirectoryException;
